code cleanup, fix bugs when exiting and re-entering level

+++ Assets/Scripts/Managers/StartManager.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
@@ -27,7 +25,7 @@ public class StartManager : MonoBehaviour
     public void OnLevelButton(int level)
     {
         // todo make loading screens
-        SceneManager.LoadScene((int)level);
+        SceneManager.LoadScene(level);
     }
     
 }
